+++ generic-client-lib/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/api/v40/ApiClient.java
@@ -3,6 +3,7 @@ package com.ptsecurity.appsec.ai.ee.utils.ci.integration.api.v40;
 import com.google.gson.reflect.TypeToken;
 import com.microsoft.signalr.HubConnection;
 import com.microsoft.signalr.HubConnectionBuilder;
+import com.ptsecurity.appsec.ai.ee.scan.result.ScanBrief;
 import com.ptsecurity.appsec.ai.ee.server.v40.auth.ApiResponse;
 import com.ptsecurity.appsec.ai.ee.server.v40.auth.api.AuthApi;
 import com.ptsecurity.appsec.ai.ee.server.v40.auth.model.AuthResultModel;
@@ -145,6 +146,11 @@ public class ApiClient extends AbstractApiClient {
         }
     }
 
+    @Override
+    public ScanBrief.ApiVersion getApiVersion() {
+        return ScanBrief.ApiVersion.V40;
+    }
+
     public JwtResponse authenticate() throws GenericException {
         @NonNull
         ApiResponse<AuthResultModel> jwtResponse;
@@ -255,6 +261,8 @@ public class ApiClient extends AbstractApiClient {
         connection.on("ScanStarted", (data) -> {
             if (!projectId.equals(data.getResult().getProjectId()))
                 log.trace("Skip ScanStarted event as its projectId != {}", projectId);
+            else if (!scanResultId.equals(data.getResult().getId()))
+                log.trace("Skip ScanStarted event as its scanResultId != {}", scanResultId);
             else {
                 if (null != console)
                     console.info("Scan started. Project id: %s, scan result id: %s", data.getResult().getProjectId(), data.getResult().getId());
@@ -312,6 +320,8 @@ public class ApiClient extends AbstractApiClient {
         connection.on("ScanCompleted", (data) -> {
             if (!projectId.equals(data.getResult().getProjectId()))
                 log.trace("Skip ScanCompleted event as its projectId != {}", projectId);
+            else if (!scanResultId.equals(data.getResult().getId()))
+                log.trace("Skip ScanCompleted event as its scanResultId != {}", scanResultId);
             else
                 queue.add(Stage.DONE);
             log.trace(data.toString());