+++ generic-client-lib/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/api/v40/converters/IssuesConverter.java
@@ -18,8 +18,7 @@ import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 
 import javax.xml.bind.DatatypeConverter;
-import java.io.InputStream;
-import java.io.InputStreamReader;
+import java.io.*;
 import java.lang.management.ManagementFactory;
 import java.lang.management.MemoryUsage;
 import java.nio.charset.StandardCharsets;
@@ -185,19 +184,20 @@ public class IssuesConverter {
     /**
      * Method converts PT AI v.3.6 API scan result and issues model pair to API version independent scan result
      * @param scanResult PT AI v.3.6 API scan result that contains scan statistic
-     * @param modelStreams PT AI v.3.6 API scan issues list with detailed information about vulnerabilities found
+     * @param modelFiles PT AI v.3.6 API scan issues list with detailed information about vulnerabilities found
      * @param scanSettings PT AI v.3.6 API scan settings
      * @return PT AI API version independent scan results instance
      */
     public static ScanResult convert(
             @NonNull final String projectName,
             @NonNull final com.ptsecurity.appsec.ai.ee.server.v40.legacy.model.ScanResult scanResult,
-            @NonNull final Map<Reports.Locale, InputStream> modelStreams,
+            @NonNull final Map<Reports.Locale, File> modelFiles,
             @NonNull final V40ScanSettings scanSettings,
             @NonNull final String ptaiUrl,
             @NonNull final Map<ServerVersionTasks.Component, String> versions) {
         ScanResult res = new ScanResult();
         convertInto(projectName, scanResult, scanSettings, versions, res);
+        res.setApiVersion(ScanBrief.ApiVersion.V40);
         res.setPtaiServerUrl(ptaiUrl);
 
         // As there's no localization in metadatas, use english model as source
@@ -207,8 +207,8 @@ public class IssuesConverter {
         // At this point we have ScanResult that is initialized with vulnerability list. But these
         // vulnerabilities have titleId field that points nowhere. So we need to create localized
         // descriptions for all of them
-        for (Reports.Locale locale : modelStreams.keySet()) {
-            IssuesModel localizedModel = parseIssuesModelStream(modelStreams.get(locale));
+        for (Reports.Locale locale : modelFiles.keySet()) {
+            IssuesModel localizedModel = parseIssuesModelStream(modelFiles.get(locale));
             // Save first model as a metadata source. As metadata is not i18n-ed, there's
             // no difference what locale will be used
             if (null == model) model = localizedModel;
@@ -637,13 +637,13 @@ public class IssuesConverter {
      * @return Deserialized IssuesModel instance
      */
     @SneakyThrows
-    protected static IssuesModel parseIssuesModelStream(@NonNull final InputStream data) {
+    protected static IssuesModel parseIssuesModelStream(@NonNull final File data) {
         JSON parser = new JSON();
         MemoryUsage usage = ManagementFactory.getMemoryMXBean().getHeapMemoryUsage();
-        try {
+        try (Reader reader = new FileReader(data)) {
             log.debug("JVM heap memory use before parse {} / {}", FileCollector.bytesToString(usage.getUsed()), FileCollector.bytesToString(usage.getMax()));
             log.debug("Parse started at {}", Instant.now());
-            IssuesModel res = parser.getGson().fromJson(new InputStreamReader(data, StandardCharsets.UTF_8), IssuesModel.class);
+            IssuesModel res = parser.getGson().fromJson(reader, IssuesModel.class);
             log.debug("Parse finished at {}", Instant.now());
             log.debug("JVM heap memory use after parse {} / {}", FileCollector.bytesToString(usage.getUsed()), FileCollector.bytesToString(usage.getMax()));
             return res;
@@ -691,6 +691,7 @@ public class IssuesConverter {
             @NonNull final Map<ServerVersionTasks.Component, String> versions) {
         ScanBrief res = new ScanBrief();
         convertInto(projectName, scanResult, scanSettings, versions, res);
+        res.setApiVersion(ScanBrief.ApiVersion.V40);
         res.setPtaiServerUrl(ptaiUrl);
         return res;
     }