+++ CHANGELOG.md
@@ -178,3 +178,4 @@
 + [Feature] As sometimes notifications service connection get lost, polling thread added to generic AST task
 + [ToDo] Remove PDF report generation as PT AI 4.0 REST API no supports it more
 + [ToDo] Remove XML / JSON report generation as PT AI 4.0 REST API no supports it more
++ [Fix] Temporal file cleanup fixed